Wrap binary op lhs/rhs in parentheses
This makes the SQL behave more like it looks in Swift.

`.binary(lhs, .ILIKE, rhs)` looks like lhs and rhs will always be
computed as `(lhs) ILIKE (rhs)`, which is not always the case.

This will only have an effect in advanced queries with precedence that's
hard to follow, but is the only simplest solution, seeing that Vapor/SQL
has no option to wrap arbitrary expressions in parentheses.